Clean the submission queue on enqueue of each track
On queueing a track, examine the head of the queue. If the track in
the head of the queue, that is, the next song to be submitted, was
played for less than 240 seconds or half its length, remove it from
the queue.

With this, only songs that have been played for at least the half their
duration will be submitted.